+++ framework/include/vasm/codegen/Binary.h
@@ -18,10 +18,10 @@ namespace codegen
     public:
         BinaryFormat();
 
-        void write(unsigned char  data, Section section) override;
-        void write(unsigned short data, Section section) override;
-        void write(unsigned int   data, Section section) override;
-        void write(unsigned long  data, Section section) override;
+        void write(unsigned char      data, Section section) override;
+        void write(unsigned short     data, Section section) override;
+        void write(unsigned int       data, Section section) override;
+        void write(unsigned long long data, Section section) override;
 
         size_t getPosition(Section section) override;
         size_t getSectionStart(Section section) override;
